Moving route_summary_cache implementation into header.
